python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
bcrypt==4.0.1